+++ src/Controller/UserController.php
@@ -31,9 +31,12 @@ class UserController extends AbstractController
     /**
      * @Route("/membres/ajout", name="app_user_member_new", methods={"GET", "POST"})
      */
-    public function newMember(Request $request, UserRepository $userRepository, UserPasswordHasherInterface $passwordHasher, 
-    SendEmail $sendEmail): Response
-    {
+    public function newMember(
+        Request $request,
+        UserRepository $userRepository,
+        UserPasswordHasherInterface $passwordHasher,
+        SendEmail $sendEmail
+    ): Response {
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -44,7 +47,7 @@ class UserController extends AbstractController
             $hashedPassword = $passwordHasher->hashPassword($user, $user->getPassword());
             // On l'écrase dans le $user
             $user->setPassword($hashedPassword);
-            
+
             $userRepository->add($user, true);
 
             // Send email with SendEmail service
@@ -66,6 +69,9 @@ class UserController extends AbstractController
      */
     public function showMember(User $user): Response
     {
+        if ($user === null) {
+            throw $this->createNotFoundException('Ce membre est introuvable');
+        }
         return $this->render('user/member/show.html.twig', [
             'user' => $user,
         ]);
@@ -76,6 +82,10 @@ class UserController extends AbstractController
      */
     public function editMember(Request $request, User $user, UserRepository $userRepository, UserPasswordHasherInterface $passwordHasher): Response
     {
+        if ($user === null) {
+            throw $this->createNotFoundException('Ce membre est introuvable');
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -109,7 +119,11 @@ class UserController extends AbstractController
      */
     public function deleteMember(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        if ($user === null) {
+            throw $this->createNotFoundException('Ce membre est introuvable');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
             $this->addFlash('success', 'Utilisateur supprimé.');
         }
@@ -142,7 +156,7 @@ class UserController extends AbstractController
             $hashedPassword = $passwordHasher->hashPassword($user, $user->getPassword());
             // On l'écrase dans le $user
             $user->setPassword($hashedPassword);
-            
+
             $userRepository->add($user, true);
 
             // Send email with SendEmail service
@@ -164,6 +178,10 @@ class UserController extends AbstractController
      */
     public function showBackOffice(User $user): Response
     {
+        if ($user === null) {
+            throw $this->createNotFoundException('Cet utilisateur est introuvable');
+        }
+
         return $this->render('user/back-office/show.html.twig', [
             'user' => $user,
         ]);
@@ -174,6 +192,10 @@ class UserController extends AbstractController
      */
     public function editBackOffice(Request $request, User $user, UserRepository $userRepository, UserPasswordHasherInterface $passwordHasher): Response
     {
+        if ($user === null) {
+            throw $this->createNotFoundException('Cet utilisateur est introuvable');
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -207,7 +229,11 @@ class UserController extends AbstractController
      */
     public function deleteBackOffice(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        if ($user === null) {
+            throw $this->createNotFoundException('Cet utilisateur est introuvable');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
             $this->addFlash('success', 'Utilisateur supprimé.');
         }
